# Improve flat_mutation_reader::consume_pausable