--- a/flat_mutation_reader.hh
+++ b/flat_mutation_reader.hh
@@ -143,11 +143,11 @@ public:
 
                 if constexpr (std::is_same_v<future<stop_iteration>, decltype(consumer(pop_mutation_fragment()))>) {
                     return consumer(pop_mutation_fragment());
+                } else {
+                    auto result = stop_iteration::no;
+                    while ((result = consumer(pop_mutation_fragment())) != stop_iteration::yes && !is_buffer_empty() && !need_preempt()) {}
+                    return make_ready_future<stop_iteration>(result);
                 }
-
-                return futurize_invoke([&consumer, mf = pop_mutation_fragment()] () mutable {
-                    return consumer(std::move(mf));
-                });
             });
         }