@haaawk 2021-03-26
@bhalevy you are the author of the commit that introduced the allocation in consume_pausable that this patchset is removing. I wasn't able to figure out why the allocation was needed but maybe I'm missing something. Please comment.