@bhalevy 2021-03-27
I see quite similar results with this patch on top of yours:
```diff
index afe02262d..5ae023ef6 100644
--- a/flat_mutation_reader.hh
+++ b/flat_mutation_reader.hh
@@ -132,8 +132,8 @@ class flat_mutation_reader final {
         // Next call will start from the next mutation_fragment in the stream.
         future<> consume_pausable(Consumer consumer, db::timeout_clock::time_point timeout) {
             return repeat([this, consumer = std::move(consumer), timeout] () mutable {
-                if (is_buffer_empty()) {
-                    if (is_end_of_stream()) {
+                if (is_buffer_empty()) [[unlikely]] {
+                    if (is_end_of_stream()) [[unlikely]] {
                         return make_ready_future<stop_iteration>(stop_iteration::yes);
                     }
                     return fill_buffer(timeout).then([] {
@@ -144,10 +144,7 @@ class flat_mutation_reader final {
                 if constexpr (std::is_same_v<future<stop_iteration>, decltype(consumer(pop_mutation_fragment()))>) {
                     return consumer(pop_mutation_fragment());
                 } else {
-                    auto result = stop_iteration::no;
-                    while ((result = consumer(pop_mutation_fragment())) != stop_iteration::yes && !is_buffer_empty() && !need_preempt())
-                        ;
-                    return make_ready_future<stop_iteration>(result);
+                    return make_ready_future<stop_iteration>(consumer(pop_mutation_fragment()));
                 }
             });
         }
```