@haaawk 2021-03-29
Pushed v2 with:
1. adding {} to while loop
2. adding benchmarks for large partitions