@haaawk 2021-03-29
@bhalevy on my machine the initial version is consistently faster:

Initial:
```
./build/release/test/perf/perf_mutation_readers --random-seed 3549335083
single run iterations:    0
single run duration:      1.000s
number of runs:           5
number of cores:          16
random seed:              3549335083

test                                      iterations      median         mad         min         max
combined.one_row                             1386697   107.638ns     0.106ns   107.532ns   108.019ns
combined.single_active                          7512    88.161us    68.463ns    87.994us    88.229us
combined.many_overlapping                        950   865.017us    82.432ns   864.850us   865.792us
combined.disjoint_interleaved                   7293    82.358us    41.170ns    82.276us    82.414us
combined.disjoint_ranges                        7327    82.094us    30.740ns    82.063us    82.211us
combined.overlapping_partitions_disjoint_rows        5558   121.429us   124.593ns   121.263us   121.580us
clustering_combined.ranges_generic           1952104   211.912ns     0.073ns   211.813ns   212.156ns
clustering_combined.ranges_specialized       2074996   185.183ns     0.043ns   185.140ns   185.266ns
memtable.one_partition_one_row                642900     1.556us     0.851ns     1.555us     1.557us
memtable.one_partition_many_rows               80477    12.512us    57.675ns    12.449us    12.619us
memtable.many_partitions_one_row               40992    24.383us    12.540ns    24.370us    24.581us
memtable.many_partitions_many_rows              3191   310.917us    67.333ns   310.850us   312.802us
```

Alternative:
```
./build/release/test/perf/perf_mutation_readers --random-seed 3549335083
single run iterations:    0
single run duration:      1.000s
number of runs:           5
number of cores:          16
random seed:              3549335083

test                                      iterations      median         mad         min         max
combined.one_row                             1327345   116.211ns     0.029ns   116.160ns   116.250ns
combined.single_active                          7377    89.650us     5.945ns    89.595us    89.656us
combined.many_overlapping                        950   859.323us   191.607ns   859.037us   859.609us
combined.disjoint_interleaved                   7166    83.762us    19.002ns    83.743us    83.855us
combined.disjoint_ranges                        7174    83.817us   159.346ns    83.619us    84.016us
combined.overlapping_partitions_disjoint_rows        5506   124.093us   133.571ns   123.842us   124.602us
clustering_combined.ranges_generic           1929620   216.821ns     0.049ns   216.773ns   216.969ns
clustering_combined.ranges_specialized       2056824   187.066ns     0.112ns   186.897ns   187.252ns
memtable.one_partition_one_row                637220     1.568us     0.496ns     1.567us     1.569us
memtable.one_partition_many_rows               78089    12.793us     2.381ns    12.791us    12.806us
memtable.many_partitions_one_row               40462    24.698us     4.539ns    24.690us    24.703us
```

I also added benchmarks for large partitions that fill flat mutation reader's buffer multiple times. Here are the results:

Baseline:
```
memtable.one_large_partition                     529     1.893ms     2.940us     1.890ms     1.897ms
memtable.many_large_partitions                    20    50.053ms    25.283us    50.028ms    50.174ms
```

Initial version:
```
memtable.one_large_partition                     563     1.779ms   126.377ns     1.778ms     1.779ms
memtable.many_large_partitions                    22    47.023ms    12.010us    47.002ms    47.237ms

```

Alternative version:
```
memtable.one_large_partition                     543     1.839ms   594.884ns     1.838ms     1.840ms
memtable.many_large_partitions                    21    48.646ms    47.701us    48.599ms    48.729ms
```

As you can see, when there are many fragments then the initial version is 6% faster than baseline and 3.2-3.3% faster than the alternative version.